% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem.test-package.R
\docType{package}
\name{golem.test-package}
\alias{golem.test}
\alias{golem.test-package}
\title{golem.test: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
